feat: Refactor authentication handling by removing useAuthCheck hook and implementing logic directly in DashboardLayout
