+++ src/app/dashboard/layout.tsx
@@ -3,19 +3,30 @@
 import type React from "react";
 import BottomNavigation from "@/components/bottomNavigation";
 import { useStore } from "@/lib/store";
-import { useAuthCheck } from "@/lib/hooks/useAuthCheck";
+import { useEffect } from "react";
+import { getAuthTokenFromStorage, getIsExistingUserFromStorage } from "@/lib/auth";
+import { useRouter } from "next/navigation";
 
 export default function DashboardLayout({ children }: { children: React.ReactNode }) {
-  const { isExistingUser } = useStore();
-  
-  // Check authentication and redirect if not authenticated
-  useAuthCheck('/onboard');
-  
-  return (
-    <div className="mx-auto min-h-screen max-w-md p-5">
-      {children}
-      {/* Only show bottom navigation for registered users */}
-      {isExistingUser && <BottomNavigation />}
-    </div>
-  );
+    const router = useRouter();
+    const { isExistingUser, setIsExistingUser } = useStore();
+
+    useEffect(() => {
+        const token = getAuthTokenFromStorage();
+        const isExisting = getIsExistingUserFromStorage();
+
+        if (token) {
+            setIsExistingUser(isExisting);
+        } else {
+            router.push("/onboard");
+        }
+    }, [router, setIsExistingUser]);
+
+    return (
+        <div className="mx-auto min-h-screen max-w-md p-5">
+            {children}
+            {/* Only show bottom navigation for registered users */}
+            {isExistingUser && <BottomNavigation />}
+        </div>
+    );
 }
